pyproject.toml
[project]
name = "walmart-stock-prediction"
version = "0.1.0"
description = "Walmart stock prediction using ELT pipeline"
authors = [
    {name = "Salman Ibrahim",email = "SalmanIbrahim521@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas (>=2.3.2,<3.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "xgboost (>=3.0.4,<4.0.0)",
    "scikit-learn (>=1.7.1,<2.0.0)",
    "matplotlib (>=3.10.6,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
